Changed for loop in drawPoints() to fix error
1. The for loop length in drawPoints() was being determined by
'filteredArray.length' before.

This array length would not get updated when a state filter was chosen
from the list, which would cause the for loop to throw an error when the
loop would run more times than there were points in the 'mapPoints'
array.

Having the for loop length be determined by 'mapPoint.length' fixed this
error that would appear when filtering states.