Added Positioning and Zoom Changes to JS, included sass-cache in gitignore
1. took out zoom controls since this will only be viewed on a mobile
screen.
2. added to point click listener function...now map will center on the
clicked point and adjust zoom depending on its current zoom level
3. added 'filterPosition' function that moves to a point that is the
average of the lat and long of the points left in the mappoints array
after filter. also adjusts zoom level if its a state and goes back to
the default position and zoom level if all is selected in the filter.