Big JS Changes, Transition to SCSS Styles
1. Moved Js styles to SCSS where applicable
2. Added JS functions to change color and center point on click
3. on further zooms point will be offset from center a set amount of
pixels
4. infodiv will come from top and leave bottom half of map exposed
5. infodiv and applicable child elements are added to the dom from the
start in the index.html and display by changing css opacity and z-index
values
6. cleaned up close info button (still needs more work)
7. changed display of filter div...(needs more work, want to make into
filter display button.
8. map now repositions on filter selection and zooms accordingly (also
goes to default positioning and zoom level on selection of "all" in the
filter options
9. added selected icon which is switched for the default icon when a
point is clicked and that point is zoomed into and the info for that
point is displayed.
10. added event listeners to the main map object to hide displayed info
and return points to default icon